changed naming, code refactoring
