+++ encryption/des_ecb.go
@@ -3,11 +3,39 @@ package encryption
 import (
 	"crypto/cipher"
 	"crypto/des"
+	"errors"
 	"fmt"
+	"strconv"
 )
 
 type DesECB struct {
 	cipherBlock cipher.Block
+}
+
+func NewTripleDesECB(key []byte) (*DesECB, error) {
+	var tripleDESKey []byte
+
+	k := len(key)
+	switch k {
+	default:
+		return nil, errors.New("creating cipher: invalid key size " + strconv.Itoa(k))
+	case 16:
+		tripleDESKey = append(tripleDESKey, key[:16]...)
+		tripleDESKey = append(tripleDESKey, key[:8]...)
+		break
+	case 24:
+		tripleDESKey = append(tripleDESKey, key...)
+		break
+	}
+
+	cipher, err := des.NewTripleDESCipher(tripleDESKey)
+	if err != nil {
+		return nil, fmt.Errorf("creating cipher: %w", err)
+	}
+
+	return &DesECB{
+		cipherBlock: cipher,
+	}, nil
 }
 
 func NewDesECB(key []byte) (*DesECB, error) {
